+++ shorts-api/src/main/kotlin/com/mashup/shorts/domain/newscard/NewsCardController.kt
@@ -1,6 +1,7 @@
 package com.mashup.shorts.domain.newscard
 
 import org.springframework.http.HttpStatus.OK
+import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
@@ -10,7 +11,9 @@ import org.springframework.web.bind.annotation.RestController
 import com.mashup.shorts.common.response.ApiResponse
 import com.mashup.shorts.common.response.ApiResponse.Companion.success
 import com.mashup.shorts.domain.newscard.dto.NewsCardResponseForm.LoadAllDetailNewsInNewsCard
+import jakarta.validation.constraints.Min
 
+@Validated
 @RestController
 @RequestMapping("/v1/news-card")
 class NewsCardController(
@@ -19,15 +22,15 @@ class NewsCardController(
 
     /**
     뉴스 카드 내의 모든 뉴스를 조회한다.
-    Param : 뉴스 카드 ID, 커서 ID
+    Param : 뉴스 카드 ID, 커서 ID, 사이즈
     Return : MutableList<LoadAllDetailNewsInNewsCard>
      */
     @ResponseStatus(OK)
     @GetMapping("/{newsCardId}")
     fun loadDetailNewsInNewsCard(
-        @PathVariable newsCardId: Long,
-        @RequestParam cursorId: Long,
-        @RequestParam size: Int,
+        @PathVariable @Min(1) newsCardId: Long,
+        @RequestParam @Min(0) cursorId: Long,
+        @RequestParam @Min(1) size: Int,
     ): ApiResponse<MutableList<LoadAllDetailNewsInNewsCard>> {
         return success(
             OK,
